requirements.txt
requests
BeautifulSoup4
Django
djangorestframework
django-filters
